src/main/java/co/edu/ff/orders/product/domain/BasePrice.java
package co.edu.ff.orders.product.domain;

import co.edu.ff.orders.common.Preconditions;
import co.edu.ff.orders.serialization.NumberSerializable;
import lombok.Value;
import org.apache.commons.lang3.StringUtils;

import java.math.BigDecimal;

@Value(staticConstructor = "of")
public class BasePrice  implements NumberSerializable {
    public static BasePrice fromNumber(Number number) {
        return new BasePrice(BigDecimal.valueOf(number.doubleValue()));
    }
    BigDecimal value;

    private BasePrice(BigDecimal value){
        Preconditions.checkNotNull(value);
        Preconditions.checkArgument(value.compareTo(BigDecimal.ZERO) >= 0);
        this.value = value;
    }


    @Override
    public BigDecimal valueOf() {
        return value;
    }
}
